task detail finish refactor and finish bug
