+++ monitool/Views/Onboarding/App/AppOnboardingView.swift
@@ -125,3 +125,13 @@ struct AppOnboardingView: View{
         }
     }
 }
+
+struct AppOnboardingView_Preview: PreviewProvider {
+    static var previews: some View {
+        AppOnboardingView()
+            .previewDevice("iPad Pro (12.9-inch) (5th generation)")
+            .previewLayout(.fixed(width: 1112, height: 834))
+        AppOnboardingView()
+            .previewDevice("iPhone 12")
+    }
+}
