+++ monitool/Views/Task Detail View/Component/ReviseButton.swift
@@ -1,0 +1,20 @@
+//
+//  ReviseButton.swift
+//  monitool
+//
+//  Created by Mac-albert on 08/08/21.
+//
+
+import SwiftUI
+
+struct ReviseButton: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct ReviseButton_Previews: PreviewProvider {
+    static var previews: some View {
+        ReviseButton()
+    }
+}
