+++ monitool/Views/Task Detail View/Today List/TodayListView+ViewBuilders.swift
@@ -38,7 +38,7 @@ extension TodayListView {
 					.font(.system(size: 20, weight: .bold))
 					.frame(minWidth: 100, maxWidth: .infinity, minHeight: 20, maxHeight: 24, alignment: .leading)
 					.foregroundColor(Color(hex: "898989"))
-
+                
 				ProofOfWork(image: "kucing2", date: "p", metricSize: matric)
 					.frame(width: matric.size.width * 0.75, height: matric.size.width * 0.75)
 					.padding(.vertical, 10)
@@ -47,8 +47,8 @@ extension TodayListView {
 						RoundedRectangle(cornerRadius: 5)
 							.stroke(Color(hex: "4EB0AB"), lineWidth: 1)
 					)
-
-				if role.isOwner {
+				
+                if role.isOwner {
 					CustomText(title: "PIC: ", content: todayListViewModel.pic?.name)
 					CustomText(title: "Notes: ", content: todayListViewModel.task.notes)
 				} else {
@@ -122,7 +122,7 @@ extension TodayListView {
 			TextField("Add notes here", text: $todayListViewModel.notesText)
 				.padding()
 				.frame(height: 120, alignment: .top)
-				.background(AppColor.accent.brightness(0.65))
+				.background(Color(hex: "F0F9F8"))
 				.modifier(RoundedEdge(width: 2, color: AppColor.accent, cornerRadius: 8))
 		}
 	}
